util-linux: update to 2.32

+++ user/util-linux/util-linux-2.3X.X/tests/run.sh
@@ -27,8 +27,21 @@ paraller_jobs=1
 
 function num_cpus()
 {
-	if lscpu -p &>/dev/null; then
-		lscpu -p | grep -cv '^#'
+	local num
+
+	# coreutils
+	if num=$(nproc --all 2>/dev/null); then
+		:
+	# BSD, OSX
+	elif num=$(sysctl -n hw.ncpu 2>/dev/null); then
+		:
+	else
+		num=$(grep -c "^processor" /proc/cpuinfo 2>/dev/null)
+	fi
+
+	# translate garbage output to "1"
+	if test "$num" -gt "0" 2>/dev/null ;then
+		echo "$num"
 	else
 		echo 1
 	fi
@@ -45,7 +58,9 @@ while [ -n "$1" ]; do
 	case "$1" in
 	--force |\
 	--fake |\
-	--memcheck |\
+	--memcheck-valgrind |\
+	--memcheck-asan |\
+	--nolocks |\
 	--show-diff |\
 	--verbose  |\
 	--skip-loopdevs |\
@@ -67,6 +82,10 @@ while [ -n "$1" ]; do
 		;;
 	--parallel=*)
 		paraller_jobs="${1##--parallel=}"
+		if ! [ "$paraller_jobs" -ge 0 2>/dev/null ]; then
+			echo "invalid argument '$paraller_jobs' for --parallel="
+			exit 1
+		fi
 		;;
 	--parallel)
 		paraller_jobs=$(num_cpus)
@@ -79,16 +98,18 @@ while [ -n "$1" ]; do
 		echo "Usage: "
 		echo "  $(basename $0) [options] [<component> ...]"
 		echo "Options:"
-		echo "  --force           execute demanding tests"
-		echo "  --fake            do not run, setup tests only"
-		echo "  --memcheck        run with valgrind"
-		echo "  --verbose         verbose mode"
-		echo "  --show-diff       show diff from failed tests"
-		echo "  --nonroot         ignore test suite if user is root"
-		echo "  --srcdir=<path>   autotools top source directory"
-		echo "  --builddir=<path> autotools top build directory"
-		echo "  --parallel=<num>  number of parallel test jobs, default: num cpus"
-		echo "  --exclude=<list>  exclude tests by list '<utilname>/<testname> ..'"
+		echo "  --force              execute demanding tests"
+		echo "  --fake               do not run, setup tests only"
+		echo "  --memcheck-valgrind  run with valgrind"
+		echo "  --memcheck-asan      enable ASAN (requires ./configure --enable-asan)"
+		echo "  --nolocks            don't use flock to lock resources"
+		echo "  --verbose            verbose mode"
+		echo "  --show-diff          show diff from failed tests"
+		echo "  --nonroot            ignore test suite if user is root"
+		echo "  --srcdir=<path>      autotools top source directory"
+		echo "  --builddir=<path>    autotools top build directory"
+		echo "  --parallel=<num>     number of parallel test jobs, default: num cpus"
+		echo "  --exclude=<list>     exclude tests by list '<utilname>/<testname> ..'"
 		echo
 		exit 1
 		;;
@@ -167,8 +188,10 @@ printf "%13s: %-30s    " "kernel" "$(uname -r)"
 echo
 echo
 
-if [ $paraller_jobs -gt 1 ]; then
-	echo "              Executing the tests in parallel ($paraller_jobs jobs)    "
+if [ "$paraller_jobs" -ne 1 ]; then
+	tmp=$paraller_jobs
+	[ "$paraller_jobs" -eq 0 ] && tmp=infinite
+	echo "              Executing the tests in parallel ($tmp jobs)    "
 	echo
 	OPTS="$OPTS --parallel"
 fi
@@ -179,6 +202,10 @@ printf "%s\n" ${comps[*]} |
 	sort |
 	xargs -I '{}' -P $paraller_jobs -n 1 bash -c "'{}' \"$OPTS\" ||
 		echo 1 >> $top_builddir/tests/failures"
+if [ $? != 0 ]; then
+	echo "xargs error" >&2
+	exit 1
+fi
 declare -a fail_file
 fail_file=( $( < $top_builddir/tests/failures ) ) || exit 1
 rm -f $top_builddir/tests/failures
